+++ .github/workflows/on-pr.yml
@@ -1,6 +1,6 @@
-name: Run all tests
+name: Run tests for PR
 on: [pull_request]
 
 jobs:
   run-tests:
-    uses: ./.github/workflows/tests.yml
+    uses: ./.github/workflows/test.yml
